fix : [백] delete service interface
