+++ src/businessservice/src/test/java/com/eventty/businessservice/presentation/EventControllerTest.java
@@ -3,7 +3,7 @@ package com.eventty.businessservice.presentation;
 import com.eventty.businessservice.domains.event.application.dto.request.EventCreateRequestDTO;
 import com.eventty.businessservice.domains.event.application.dto.response.EventWithTicketsFindByIdResponseDTO;
 import com.eventty.businessservice.domains.event.application.dto.response.EventBasicFindAllResponseDTO;
-import com.eventty.businessservice.domains.event.application.Facade.EventServiceImpl;
+import com.eventty.businessservice.domains.event.application.Facade.EventService;
 import com.eventty.businessservice.domains.event.presentation.EventController;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
@@ -31,7 +31,7 @@ public class EventControllerTest {
     private ObjectMapper objectMapper;
 
     @MockBean
-    private EventServiceImpl eventServiceImpl;
+    private EventService eventService;
 
 //    @Test
 //    @DisplayName("특정 행사 조회 테스트")
@@ -57,7 +57,7 @@ public class EventControllerTest {
     public void findAllEventsTest() throws Exception {
         // Given
         List<EventBasicFindAllResponseDTO> mockEventList = createEventRespnseDTOList(3L);
-        when(eventServiceImpl.findAllEvents()).thenReturn(mockEventList);
+        when(eventService.findAllEvents()).thenReturn(mockEventList);
 
         // When & Then
         mockMvc.perform(get("/events"))
@@ -67,7 +67,7 @@ public class EventControllerTest {
                 .andExpect(jsonPath("$.successResponseDTO.data").isArray())
                 .andExpect(jsonPath("$.successResponseDTO.data.length()").value(mockEventList.size()));
 
-        verify(eventServiceImpl, times(1)).findAllEvents();
+        verify(eventService, times(1)).findAllEvents();
     }
 
     @Test
@@ -76,7 +76,7 @@ public class EventControllerTest {
         // Given
         Long newEventId = 10L;
         EventCreateRequestDTO eventCreateRequestDTO = createEventFullCreateRequestDTO();
-        when(eventServiceImpl.createEvent(eventCreateRequestDTO)).thenReturn(newEventId);
+        when(eventService.createEvent(eventCreateRequestDTO)).thenReturn(newEventId);
 
         // When & Then
         mockMvc.perform(post("/events")
@@ -92,7 +92,7 @@ public class EventControllerTest {
     public void deleteEventTest() throws Exception {
         // Given
         Long eventId = 1L;
-        when(eventServiceImpl.deleteEvent(eventId)).thenReturn(eventId);
+        when(eventService.deleteEvent(eventId)).thenReturn(eventId);
 
         // When & Then
         mockMvc.perform(delete("/events/{eventId}", eventId))
@@ -100,7 +100,7 @@ public class EventControllerTest {
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                 .andExpect(jsonPath("$.success").value(true));
 
-        verify(eventServiceImpl, times(1)).deleteEvent(eventId);
+        verify(eventService, times(1)).deleteEvent(eventId);
     }
 
     @Test
@@ -108,7 +108,7 @@ public class EventControllerTest {
     public void findEventsByCategoryTest() throws Exception {
         // Given
         Long categoryId = 3L;
-        when(eventServiceImpl.findEventsByCategory(categoryId)).thenReturn(createEventRespnseDTOList(5L));
+        when(eventService.findEventsByCategory(categoryId)).thenReturn(createEventRespnseDTOList(5L));
 
         // When & Then
         mockMvc.perform(get("/events/category/{categoryId}", categoryId))
@@ -116,7 +116,7 @@ public class EventControllerTest {
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                 .andExpect(jsonPath("$.success").value(true));
 
-        verify(eventServiceImpl, times(1)).findEventsByCategory(categoryId);
+        verify(eventService, times(1)).findEventsByCategory(categoryId);
     }
 
     private static EventCreateRequestDTO createEventFullCreateRequestDTO() {
